Rework structure
